feat(#693): Duplicate Code: scrapling - massive mock boilerplate duplication in venv-setup-scrapling.test.ts
